refactor(invite code monitor table): extract invite code monitor table to a dedicated component
